And the conflicting txt module
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@1537 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ modules/txt/operations/parser_lines.h
@@ -43,8 +43,8 @@ helpLine: Parse a txt file emiting every line as an independent string ( return 
 		void run( char *data , size_t length , samson::KVWriter *writer )
 		{
 		  
-			int pos_begin=0;
-			int pos_end;
+			size_t pos_begin=0;
+			size_t pos_end;
 
 			while( pos_begin < length )
 			{
